crawler/requirements.txt
selenium
webdriver_manager
psycopg2
Pillow
pandas
openpyxl

